+++ networkmanager/src/config/ip6.rs
@@ -32,7 +32,7 @@ impl<'a> Ipv6Config<'a> {
 					let address_str = address_str.downcast_ref::<zvariant::Str>()?;
 					Ipv6Addr::from_str(address_str).ok()?
 				};
-				let prefix = u64::try_from(map.remove("prefix")?).ok()? as usize;
+				let prefix = u32::try_from(map.remove("prefix")?).ok()?;
 				Some(AddressData { address, prefix })
 			})
 			.collect())
@@ -56,5 +56,5 @@ impl<'a> From<Ipv6ConfigProxy<'a>> for Ipv6Config<'a> {
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct AddressData {
 	pub address: Ipv6Addr,
-	pub prefix: usize,
+	pub prefix: u32,
 }
